src/requirements.txt
pyyaml
requests
deepdiff
flask
